backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
apscheduler==3.10.4
google-generativeai==0.4.0
httpx==0.27.0
slowapi==0.1.9
